requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.8.0
torchvision==0.23.0
transformers==4.56.1
pillow==11.0.0
huggingface-hub==0.34.5
fastapi[standard]==0.116.2
uvicorn[standard]==0.35.0
python-multipart==0.0.20